Ajuste campos tipo do ID

+++ src/main/java/br/com/bruno/musicas_a_p_i/repos/MusicasRepository.java
@@ -4,5 +4,5 @@ import br.com.bruno.musicas_a_p_i.domain.Musicas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface MusicasRepository extends JpaRepository<Musicas, Long> {
+public interface MusicasRepository extends JpaRepository<Musicas, Integer> {
 }
